# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Property Pro Lite RESTFUL API",
    "description": "Property Pro Lite is a platform where people can create and/or search for properties for sale or rent",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Blessing Makaraba",
      "email": "blessingmakaraba@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "host": "propertyprolite1.herokuapp.com",
  "basePath": "/api/v1",
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "scheme": "bearer",
      "description": "You are required to pass the JWT token received upon signup and signin in the headers to access secured endpoint, required input format: \"Bearer \\<token\\>\"",
      "name": "Authorization",
      "in": "header"
    }
  },
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/auth/{user_email}/reset_password": {
      "post": {
        "tags": ["Users"],
        "summary": "user can reset his/her password",
        "description": "",
        "operationId": "resetPassword",
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "user_email",
            "in": "path",
            "description": "type of property to fetch",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "body",
            "description": "old password and new password of user",
            "schema": {
              "type": "object",
              "properties": {
                "old_password": {
                  "type": "string",
                  "example": "Blesdfnnjkjk9"
                },
                "new_password": {
                  "type": "string",
                  "example": "Blesdfnjk9"
                }
              }
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "422": {
            "description": "email is invalid"
          },
          "404": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
  },
...
}